+++ lib/model/vertex.dart
@@ -100,4 +100,9 @@ class Vertex<I> {
   ///
   /// 当前节点是否有鼠标浮入
   bool get isHovered => cpn?.isHovered ?? false;
+
+  @override
+  String toString() {
+    return '$position';
+  }
 }